Simplify "compile" step in Ant build.
Summary:
In the course of modifying the Ant build, I started to wonder why there
were two `ecj` tasks in the `compile` target. I looked back in history
and found an earlier revision with a helpful rationale comment before the
first `ecj` task, that laid out a rationale which appeared to no longer
hold.

So, I deleted the first `ecj` task, discovered that the classes generated
in the `build` directory are bit-for-bit identical, and called it good. :-)

Test Plan:
Binary comparison of the `build` directory created by `ant default` before and
after the change. `.class` files should be identical. `.jar` files will be
different because they contain timestamps, but unjar them and their contents
will be identical.