C/C++: support expanding macros in flags and search paths
Summary:
This is part 1 in fixing an issue with apple libraries, where we
eagerly expand apple platform-specific paths before we actually know
which platform to target.  In this diff, we add the concept of macros
to the `CxxPlatform` object which get expanded by the C/C++ layer at
which point the platform is known.

Test Plan: added unittest