Use try-with-resources where appropriate to prevent resource leaks
Summary:
In several places Eclipse warns that resources are potentially not closed
if an exception occurs. In some of these places, try-with-resources can be
used so that the resource is properly closed on exception.

In other places using try-with-resources is not appropriate because it
will cause an underlying stream to be closed, which is not wanted.  In
those places, suppress the warning instead.

Test plan:

* `buck test`
* Check that warnings are no longer shown in Eclipse.