Make DX's Optimizer not use static variables
Summary:
[Forward Port of 55ad08e0fdc80153738008dffa8208f65fdd838a]
We don't actually need the optimizer when running in-process, so
originally I was just going to disable it, but I want all of dx's tests
to be able to run with my modifications, so I need to make it work.

The general pattern in the optimizer was:

Someone calls a static Optimizer.optimize() method.
The optimizer sets some of its static variables and calls various sub-methods to do the optimization.
Those sub-methods inspect Optimizer's static variables to get their configuration.
The new pattern is:

Someone calls new Optimizer.optimize().
That optimizer sets non-static variables.
The optimizer passes itself down to the sub-methods, which stash the pointer in a field so they can query its variables.

Test Plan: testing after merges are complete