Enforce immutability in DX's RegisterSpecSet
Summary:
[Forward Port of f5ac42bf8cafa63accd37d998f53a832edde893d]
It wasn't calling EMPTY.setImmutable().
It wasn't calling throwIfImmutable() in remove().
It wasn't directly calling throwIfImmutable() if putAll(). Currently, putAll would throw if it actually tried to add anything, but if someone tried to optimize it by replacing the call to put() with manual array optimizations, it would not throw.

Test Plan: ant dx-tests