AdbHelper: only restart adb when no device is attached
Summary:
Currently the adb server is restarted both when there is no device
attached, or when there is more than one device attached, but the user
hasn't enabled multi install mode.  Restarting the adb server helps when
adb is flaky and shows devices as offline or not attached, but it
doesn't help when there are too many devices are attached and the user
forgets to enable multi install mode.  This is usually a user error.

Disable the server restart and save a little time for the user.

Test plan:
Test buck install with different device configurations and show it does
the right thing about restarting the adb server.
Updated unit tests.