Make BsdScrubber work with sorted symbol tables
Summary:
Previously, our mach-O symbol table scrubber relied on the descriptor
array ordering matching the symbol name table ordering, and used that
to find the last symbol name table entry.  However, this doesn't appear
documented as something we can rely on, and the when using a sorted
symbol table (e.g. `ranlib -s ...` or default `ar`) this isn't the case.

This diff changes to search all descriptors to find the last entry.

Test Plan: Built a C/C++ binary on Darwin and verified it wasn't corrupted.