Fix a `FileAlreadyExistsException` on Unix Systems
Summary:
One of the downsides to me developing Buck on Windows is that I don't
actually create symlinks with Buck ever.  This hid a bug where I was
not cleaning out my scratch directory for Rust code before
generating a symlink tree (because copying a file to a place where
one exists is perfectly fine).

Test Plan: Added a test that fails on Unix systems (but not Windows!)