Introduce an EventKey to identify BuckEvents.
Summary:
Currently Start/Finished BuckEvent pairs are found by calling
isRelatedTo. Unfortunately that means we can't easily match
these pairs when looking at the data uploaded to the remote
log.
This change introduces the concept of an EventKey. The EventKey
is simply a value which should be shared by all events which
are related to each other. For most events this can be
a sequence number, however that requires the Finished event
to have access to the Started event which is not always
convenient. For these cases an alternative form of the
EventKey is availabe where it's simply the concatenation
of the significant fields in the event. The alternative
from should be used sparingly as it's a lot more expensive.

Test Plan: CI, manually inspect log output.