Add quiet option to installApk
Summary:
Sometimes when calling the installApk method of AdbHelper, it is better
if success events are kept quiet.  For example when running Android
instrumentation tests, the success events and outputs to the console
would only clutter the output, but give no additional information.
Outputs that notify the user of errors are still shown.

Test plan: Add unit tests.