Load TranslatingJavacPhaseTracer from resources
Summary:
This commit ties the last several commits together to enable tracing of `javac`
compilation phases to `BuckEventBus`.

`TranslatingJavacPhaseTracer` needs to be loaded anew in each compiler's class
loader. To accomplish this, we bundle its jar into Buck as a java resource, and
load it from there at runtime.

Test Plan:
Verify functionality:
* `git clean -fdx && buck build buck`
* Verify buck-out/log/traces/build.trace contains javac phase tracing
* `buck kill && cp buck-out/gen/programs/buck.pex . && rm -r build buck-out && ./buck.pex build buck`
* Verify buck-out/log/traces/build.trace contains javac phase tracing

Verify overhead:
Run a number of builds of a largish target with and without this diff. Observe the javac time, verify it does not increase more than a nominal amount due to the extra tracing.