StubJar: don't buffer classes in memory
Summary:
The tree set we use to order the classes in the output stub jar holds
all the classes in-memory.  For really large JARs (e.g. uber JARs) this
can consume a lot of memory.

Since the only use-cases right now operate on JARs, and since the
JARs we produce are deterministic, we don't necessary need to use a
tree set to detemrinize the output.  This diff removes the buffering
so that the generating stub jars is order preserving.

Test Plan: added unittest