Add AdbHelper.get() method
Summary:
ExecutionContext contains almost all fields needed to create an
AdbHelper.  Add the remaining necessary fields (AdbOptions and
TargetDeviceOptions), and add a AdbHelper.get() method, which only needs
the ExecutionContext and a flag whether the adb server should be
restarted on failure or not.

Test plan: Check that buck install still works as expected