buck.py perf [4/7]: Check version of Watchman to conditionally enable watchman globbing
Summary:
This is part 4 of 7 of a series of big performance fixes to
the `buck.py` parser of `BUCK` files.

In this diff, I check the version of Watchman to see if it's >= 3.6.0. If so,
I know it has the new rewritten glob parsing logic I added upstream:

https://github.com/facebook/watchman/commit/471138f2cbcea1c974a7c539a4522fa741738cf4

I repurposed the unused method `WatchmanWatcher.isWatchmanAvailable()` for
the version check.

Test Plan:
`buck build buck` with `watchman` 3.5.0, check logs, confirm watchman
version logged correctly and flag set to false.

Repeated with `watchman` 3.6.0, confirmed version logged correctly and flag
set to true.