buck.py perf [1/7]: Easy: Refactor ProjectBuildFileParser options into immutable value type
Summary:
This is part 1 of 7 of a series of big performance fixes to
the `buck.py` parser of `BUCK` files.

In this diff, I refactor the bucket of options passed to
`ProjectBuildFileParser` into a new immutable value type
`ProjectBuildFileParserOptions`, so I can more easily add another
option with a default value without touching unrelated files.

Test Plan: CI