buck.py perf [5/7]: New buck.py option --use_watchman_glob
Summary:
This is part 5 of 7 of a series of big performance fixes to
the `buck.py` parser of `BUCK` files.

In this diff, I add a new option `--use_watchman_glob` to `buck.py`. If present,
we use a new `watchman_glob()` method I implemented to use `watchman query`
instead of Python's `pathlib` to glob for files.

This results in a net improvement of about 50% in `BUCK` file parse
times for large repositories with tens or hundreds of thousands of
files.

Note that I had to update `buck`'s `.watchmanconfig`, as we had
`third-party/java/dx` excluded from Watchman from Day 1 (@jimp's diff
`Ignore directories in Watchman`). This would cause `watchman query`
to return nothing for that directory, which would cause `dx`
not to be built.

Test Plan:
`buck build largeproject` with `watchman` 3.5.0 in `$PATH`.
Confirmed project took about 60 seconds to parse `BUCK` files. in
Chrome trace, confirmed that invoking `buck.py` was the bulk of the time.

Updated `$PATH` to point to `watchman` 3.6.0. Re-ran test.
Before, it took about 60 seconds. Now, it takes about 30 seconds.