buck.py perf [6/7]: Conditionally build and use native pywatchman client
Summary:
This is part 6 of 7 of a series of big performance fixes to
the `buck.py` parser of `BUCK` files.

In this diff, I import the `pywatchman` native Watchman client from
the upstream Watchman revision
75ef2315244ce5644d57feacbdf8eb15c7daa813.

I added `ant` compilation for this Python module, storing the
result in `build/pywatchman`. (To support IntelliJ and Eclipse
developers, we fall back to the old non-Watchman path if
this module is not present.)

This client has two huge improvements over using `subprocess` to
fork a Watchman client for each query:

1) It uses a persistent client which communicates to Watchman
over a socket

2) It uses the efficient BSER encoding for queries and responses
instead of the very inefficient JSON encoding

This results in about a 33% improvement in performance of parsing
`BUCK` files with lots of globs.

Test Plan:
`buck build largeproject` with `watchman` 3.5.0 in `$PATH`.
Confirmed project took about 60 seconds to parse `BUCK` files. in
Chrome trace, confirmed that invoking `buck.py` was the bulk of the time.

Updated `$PATH` to point to `watchman` 3.6.0. Re-ran test.
Before, it took about 60 seconds. Now, it takes about 24 seconds.

Built Buck PEX with `buck build buck`. Ran it in a test repo. Confirmed it tried and failed to import `pywatchman` but continued to work otherwise.