C/C++: generate and use dependency files to avoid preprocessing
Summary:
This hooks up the C/C++ rules to use the dependency files supported
by the build engine to avoid re-preprocessing when unused headers
are changed.  Since we tend to overspecify in the UI, this can help
reduce the amount of preprocessing we do.

Addresses part of #298

Test Plan: added tests