python: add support for a fast-building in-place package variant
Summary:
Currently, the only output option for a python binary is creating
a stand-alone package which combines all the modules, resources,
and native libs found from the transitive deps.  Since this packaging
step is costly, this can add significant overhead to workflows that
remain in-tree.  This diff adds an output option for python binaries
(and tests) which avoid the packaging step, and instead generate a
symlink tree to find the various components.

Test Plan: added tests