Add an option to disable MissingSymbolsHandler
Summary:
Add a config option to skip MissingSymbolsHandler.

This handler is a helper class that waits for all builds to finish and lists missing dependencies.

The biggest problem with this logic is strange hanging at the end of failed build. It can take up to several seconds while with this option enabled it's usually around 1 second.

Another problem is that this class doesn't always work as intended even though JavacStep lists missing dependencies correctly. I usually get output like that:

  Rule //java/com/facebook/chatheads/prefs:prefs has failed to build.
  FbSharedPreferences
  FbSharedPreferencesModule
  PrefKey
  com.facebook.prefs.shared
  com.facebook.prefs.shared.FbSharedPreferences
  com.facebook.prefs.shared.FbSharedPreferencesModule
  com.facebook.prefs.shared.PrefKey
  Try adding the following deps:
  //java/com/facebook/prefs/shared:shared
  BUILD FAILED: //java/com/facebook/chatheads/prefs:prefs failed with exit code 1:
  javac

I would remove that handler at all because JavacStep already lists missing dependencies.

Test Plan: Remove some dependency from a BUCK file and try to build with this change and without it.