Fix failures with python3 as default interpreter
Summary:
On systems on which python refers to python3, building buck fails in the
python-get-prefix and in the compile-pywatchman steps.  In addition,
AndroidBinaryIntegrationTest fails in the testPreprocessorForcesReDex
test, because of a python script that is not compatible with python3.
Make the script and ant build compatible with python3.

Test plan:
buck test
ant - with python3 as default python