Pom generator
Summary:
To publish to maven central anything that has dependencies, we have to publish a pom file alongside.
Also, given that the pom files tend to contain a lot of information, that we cannot auto-populate, it's reasonable to assume that user may want to add data other than dependencies to the pom file, before publishing it to the central repository, so ideally, we need to also support updating existing `pom.xml` without overwriting it as much as possible.

Test Plan:
  buck test   --filter MainIntegrationTest test/com/facebook/buck/cli &&
  buck test   --filter PomCommandIntegrationTest test/com/facebook/buck/cli &&
  buck test   --filter XmlDomParserTest test/com/facebook/buck/util