Log only exceptional events as `SEVERE`
Summary:
This diff restores logging `SEVERE` log events to the console, but
reserves this for truly exceptional events, as it causes the super
console to freeze.  In particular, we make sure to limit normal
console events to `WARNING`s when forwarding them to the logging
framework.

Test Plan:
Ran a local build an verified compilation error in build didn't
freeze super-console