Fix watchman glob when symlinks are globbed
Summary:
When I fixed the watchman glob matching directories,
I inadvertently broke it matching symlinks (since watchman
doesn't follow symlinks).

I plan on fixing this in Watchman, but for now, this is a short-term
fix to assume symlinks which match a glob are files (and not directories).

Test Plan: buck build //path/to:lib in repo with glob matching symlink