Set the full path to buck in the Xcode script.
Summary:
Xcode doesn't inherit PATH from the terminal, so the shell script usually
ends up being unable to find buck. To compound the problem, the script will
not fail when it cannot find buck. This addresses both issues by setting -e
and writing the full path to buck in the script as detected by buck itself.

Test Plan:
CI
generate a project with --build-with-buck, build and run
change the path to buck to be incorrect, build and run, see error