Offload buffer space to BufferedReader
Summary:
Reading characters from BufferedReader is more efficient than using a
straight Reader because a non-buffered reader reads byte by byte to
determine the character set which results in multiple I/O operations.

To ensure this diff doesn't cause a BufferedReader to be wrapped in
a BufferedReader and trigger the resulting double buffered performance
hit there's a constructor specifically for BufferedReaders and a
separate one to wrap non-buffered readers in a BufferedReader.

Test Plan: - Local testing