Add tests for transitive dependencies in aars
Summary:
Aars can depend on resources that are defined in another aar, most
notably the google design library depending on resources defined in the
appcompat library.  This can cause build errors or app crashes, if the
resources are not included in the build.  Add tests to make sure we
don't regress the recent fix for such aars.

Test Plan:
Added tests
Checked a project using the design library works