Fix problem with using build target patterns in multiple queries.
Summary:
The way inputs are resolved into build targets in `buck query` was changed to support build target patterns.

However, using patterns was not working in multiple queries because the
inputs of multiple queries were formatted as fully qualified build targets.
In practice, this is not useful anymore because the new way of resolving targets does not need this.

Test Plan: integration tests