Infer source path names from `BuildTargetSourcePath`s
Summary:
The C/C++ support relies on extracting names from it's source paths
to generate output paths.  Previously, for `BuildTargetSourcePath`s,
we just used the short name of the wrapped build target.  However,
in the case of thrift, we supply several sources to the same library
generated from the same thrift compile rule, and so the source path
names would conflict.

This diff changes the name inference code to use the explicitly set
path for `BuildTargetSourcePath`s when available, so that we are
more likely to generate a unique name.

Test Plan: Built C/C++ flavor of thrift library rule.