Fix PythonInPlaceBinary rule key calculation
Summary:
In his diff `python: add support for a fast-building in-place package variant`,
@andrewjcg split `PythonBinary` into two subclasses, `PythonInPlaceBinary` and
`PythonPackagedBinary`.

Andrew made sure `PythonPackagedBinary` used `@AddToRuleKey` to calculate
the rule key correctly, but forgot to do so for `PythonInPlaceBinary`. This
meant any time a repo set `[python] package_style = inplace`, all
Python targets would have rule keys which wouldn't change even if
their inputs changed.

This adds `@AddToRuleKey` to the fields of `PythonInPlaceBinary`. I'm
not sure if the `WriteFile` should also affect the rule key or not,
but I added it just in case.

Test Plan:
Added to `.buckconfig`:

  [python]
    package_style = inplace

Ran `buck test //path/to:python-test` then changed an input file
to that test. Previously, the test would be cached and would not be
re-built or re-ran. Now, the test is not cached and is re-built and
re-ran.