Speedup finding build-time deps via the classpath macro
Summary:
Previously, the classpath macro called `getTransitiveClasspathEntries`
to find all the libraries contributing to the classpath for a
library rule and extracted the build-time deps from that.  However,
the `getTransitiveClasspathEntries` returned a multi-map which also
maintained a mapping from the `JavaLibrary` rules to the JARs they
provided.  To make the suggested deps functionality work with exported
deps, this also included a mapping from java libraries to every JAR
they pulled in via transitive exporteddeps.  Building up these mappings
makes this function really slow.

Since we only care about the actual `JavaLibrary` rules, this diff adds
a sort of fast-path function, `getTransitiveClasspathDeps` which only
maintains the set of deps, which we can use at action-graph construction
time.  This function also uses `ImmutableSet` instead of `ImmutableSortedSet`
to avoid the significant cost of sorting.  This method offers up to a 10x
speedup in workloads which use the `classpath` macro extensively.

Test Plan: Updated unittests to test new method.