python: use `Tool` abstract to manage how pex builder is added to rule keys
Summary:
We recently started, correctly, adding the PEX builder tool to the
rule key of packaged python binary rule.  However, since we add this
as a path, this uses the contents of the PEX builder tool behind the
path, which unfortuntely varies quite a bit:
- When we build Buck as PEX, we package the builder tool as a PEX
  itself.  Since PEX files are ZIP files, which include mtimes, the
  contents of the builder tool is never the same from different builds
  of the Buck PEX, even when they come from the same Buck version.
  This means deploying separate Buck PEXs for different architectures
  from the same Buck version will always generate different rule keys
  for python binary rules.
- Executing Buck directly from it's repo uses a python script as the
  builder tool, meaning rule keys for python binary rules will be
  different between PEX-based builds and in-source-based builds.

This diff switches to use the `Tool` abstraction to model the pex
builder, which gives us more control over how it's added to rule
keys.  In particular, since the default PEX builder is a packaged
component of Buck, the Buck version *should* be an adequate summary
for rule keys -- the PEX tool, afterall, is included in the Buck
repo, and so changes to it will change the Buck version.  So, switch
to using a `VersionedTool` for the default PEX builder tool, which
uses the Buck version in rule keys, rather than the hash of the tool.

Test Plan:
Verified rule keys for packages python binary rules are identical
with PEX-based and in-source builds.