[buck] Add configuration options for selecting the default gtest main function
Summary:
This diff allows you to specify the test main dependency independently of the gtest dependency and to enable or disable adding that dependency at all. By default buck will use the old behavior and the value of the dependency for the main function will be the same as that of gtest.

Test Plan:
1) `ant clean && ant test -Dtest.class=CxxTestDescriptionTest` (ant clean test fails on AppTest, which I'm not sure how to deal with)