Perf: Decouple watchman glob from buckd functionality
Summary:
Currently, the Watchman-based glob feature of Buck
is coupled to the Buck daemonizing functionality. (We don't bother
setting up the Watchman watch if we're not daemonized.)

This diff decouples daemonizing from Watchman globbing, and
enables the Watchman watch if either we're daemonized or
the user has set in `.buckconfig.local`:

  [project]
    glob_handler = watchman

This lets us greatly simplify the `Main.Daemon` inner class.
Specifically, it no longer has to set up the `WatchmanWatcher`
class. (It does need to hold on to the `EventBus` and the
`UUID` both of which `WatchmanWatcher` re-uses between
runs.)

When I fixed this, I noticed `DaemonIntegrationTest` has been
disabled for a while. This was because a new feature called
`root_restrict_files` was enabled which disallows arbitrary
watches if a specific file (say, `.git`) is not found in the
watch directory.

I fixed the test and re-enabled it. I also removed some stale
logic in `DaemonIntegrationTest.waitForChange()` which is
not needed in the Watchman-only world (Watchman already
uses a cookie file to ensure all changes are flushed).

Finally, I have no idea how the test
`DaemonIntegrationTest.whenSourceInputInvalidatedThenRebuildFails()`
ever passed, but it wasn't handling the `HumanReadableException`
which fires if the file it deleted is noticed and properly raised
as an error. So I fixed that too.

I had to update the `buck.py` launcher to pass along saved versions of `$TEMPDIR`, `$TEMP`, `$TMPDIR`, and `$TMP` since we blow the Java default temp directory away by setting `-Djava.io.tmpdir` in the Buck launcher and in `JUnitStep`. We also blow away `$TMP` in `JUnitStep` for good measure.

Test Plan:
  ant java-test -Dtest.class=DaemonIntegrationTest

Added to `.buckconfig.local`:

  [project]
    glob_handler = watchman

then ran:

  NO_BUCKD=1 buck build buck

Confirmed in `buck-out/log/buck-0.log` that `--use_watchman_glob`
and `BSER` were both enabled even though we weren't daemonized.

Ran `TMP=/path/to/dir buck test -f DaemonIntegrationTest //test/com/facebook/buck/cli:cli` and confirmed via `ls -l /path/to/dir` that temporary files were in fact created in that directory and not in `/tmp`.