New JNA class HostnameFetching to work around Java hang when DNS misconfigured
Summary:
Users reported Buck hanging when the network was misconfigured.

We tracked this down to Java's implementation of
`InetAddress.getLocalHost().getHostName()`, which hangs indefinitely
if DNS is not working.

This replaces it with a JNA call to `gethostname()` (on non-Windows systems)
or `GetComputerNameEx()` (on Windows systems).

Test Plan: Unit test included. `ant java-test -Dtest.class=HostnameFetchingTest`