Genrule: include items expanded from macros in rule key
Summary:
This is the prereq to having `Genrule`s support input-based rule keys.
Currently, items used from expanded macros aren't included in the rule
key of the genrule in any way other than via the fact that normal rule
keys include the rule keys of their deps.  This diff adds a new function
to the `MacroHandler` to allow macros to provide items that should be
added to the rule key of the using rule:

For example:
- Location macros propagate the `SourcePath` wrapping their rule.
- Executable macros propagate the `Tool` object describing the executable.
- Classpath macros propagate all the `SourcePath`s wrapping the respective
  JARs.

Test Plan: added unittests