Extend depfile rule keys to support added headers.
Summary:
Depfile rule keys currently support the case of the contents of the headers
being changed, but they always miss if the header maps themselves change.
This extends the caching build engine to filter the contributions of header
symlink trees to the rule key to support that scenario.

Test Plan:
CI
new tests