Enhance TargetNode equality to work with multiple repos
Summary:
It's entirely possible for each repo to have targets with duplicate
names, but which represent different rules. Currently when we build
the Target Graph we only use the build target was for comparison, &
this means that we're ignoring the repo the target is in. Fix that.

Test Plan: buck test --all