Disable dsym generation for test bundles
Summary:
Dsym generation is expensive both in time and space. It takes about as much
time as linking and more space than the linked binary. This becomes important
when there are many statically linked libraries.

Currently, dsym files don't seem to be used when crash reporter is reporting a
test crash (it still uses the symbols to symbolicate, but line numbers are not
reported). This makes their immediate value dubious.

In the long term, we should probably figure out how to get debug info to work
correctly. For now, it makes sense to disable this.

Test Plan:
build a test project that crashes, see that crash reporter still symbolicates.

See that dsym is not generated for the test bundle.