Avoid building the target graph as much as possible in buck query.
Summary:
Currently, the target graph is being updated when new targets are found. The target graph is immutable and the only way to build it is to generate the transitive closure of a set of targets.

In many queries, the transitive closure is not necessary to get the answer. In this diff, the target graph generation is minimized as much as possible.
The parser is now always used then resolving targets, even aliases. This way, it's possible to get the `TargetNode`s from the parser cache. Queries like
  $ buck query "//foo:bar" --output-attributes buck.type name deps

Become instantaneous because buck is no longer building the whole target graph.

Test Plan:
Integration tests.
Running `buck query` on local projects.