New class ListeningProcessExecutor
Summary:
The existing `ProcessExecutor` has a few flaws:

1) It spins up three threads per process to handle I/O
2) There's no way to wait for the process to exit with a timeout
3) It prints process output to `Console` by default

This introduces a new `ListeningProcessExecutor` which is built
on the more efficient NuProcess library (~ 90 KiB JAR):

https://github.com/brettwooldridge/NuProcess/

`NuProcess` uses a constant number of threads to manage process I/O
instead of O(N) on the number of processes. (It uses `kevent` on OS X,
`epoll` on Linux, and overlapped I/O on Win32.) The library exposes an
asynchronous callback mechanism to notify listeners when the process
starts, has I/O, or exits.

Java 7 uses a default stack size of 1 MiB, so this potentially saves
us 64 * 3 = 192 MiB of memory dedicated just to thread stack storage
when we're running 64 `clang` processes simultaneously. (If we're
running in piped compilation mode where we launch `clang` and `cpp` in
parallel, this will save us 384 MiB.)

I've contributed a number of fixes upstream to `NuProcess`, all of
which have been merged except for one:

https://github.com/brettwooldridge/NuProcess/pull/32

I included that patch in the jar I added and mentioned the
commit in `README.facebook`.

Test Plan: Unit tests included. `ant java-test -Dtest.class=ListeningProcessExecutorTest`