Add filter operator to buck query.
Summary:
Add a `filter(pattern, expression)` operator to buck query, which evaluates the
given expression and filters the result by applying the given pattern to the
targets' names.

Example:
  $ buck query "filter('\.java$', labels(srcs, deps(//foo:bar)))"
Outputs all java files used to build `//foo:bar`.

Extracted the logic of a pattern matching query operator to an abstract class
to reuse code between `filter` and `kind`.

Test Plan:
tests and running on the command line.
Published the docs locally to inpect the output.
  $ bash docs/soyweb-local.sh