Push more logic down into the DefaultFileHashCache
Summary:
Prior to this change, the `MultiProjectFileHashCache` acted as
a container of other `FileHashCache` instances and implemented
some of the logic that rightfully belonged in the caches.

We rename `MultiProjectFileHashCache` to `StackedFileHashCache`
and make it implment the Chain of Responsibility design pattern
which, in turn, makes it possible to compose them.

Test Plan: buck test --all