Allow Repositories to be chained from one-another.
Summary:
This is essential for cross-repo builds.

Landing in its own diff, without hooking it up properly so that
it's out there and so that later diffs don't bloat out too much
(which would clearly be less than optimal).

This is still highly experimental, not hooked up, and liable to
change, so not documenting just yet.

Test Plan: buck test --all