Make FullyReadOnCloseInputStream.close only read once.
Summary:
This makes the FullyReadOnCloseInputStream not attempt to read
bytes when a closed stream is closed. Otherwise we get an exception
on success because the FilterInputStream in readFetchResponse
closes the stream first and then we attempt to close the stream again
when we exit the fetchImpl method.

Test Plan: I used an integration test from another diff I'm working to simulate an end-to-end scenario.