New testutil class FakeListeningProcessExecutor
Summary:
Previously, I introduced a new class
`ListeningProcessExecutor` along with tests.

This diff introduces a fake version of `ListeningProcessExecutor`,
`FakeListeningProcessExecutor`, which we can use in tests for classes
which depend on `ListeningProcessExecutor`.

Since `ListeningProcessExecutor` has a bit more functionality than
`ProcessExecutor`, the fake is a bit more complex. I modeled it as a
simplistic state machine which can move between reading from stdin,
writing to stdout or stderr, waiting for time to elapse, or exiting.

Using `FakeListeningProcessExecutor` is similar to `FakeProcessExecutor`,
but instead of providing a map of `ProcessExecutorParams` to `FakeProcess`
instances, you now provide a map of `ProcessExecutorParams` to
`Collection<FakeListeningProcessState>` to represent the state machine
of the fake process.

For now, the state machine can only move forward; it can't loop
between states.

Test Plan: Unit tests included. `ant java-test -Dtest.class=FakeListeningProcessExecutorTest`