Upgrade to latest upstream pywatchman and add to PEX
Summary:
Previously, we only supported `pywatchman` in the Ant
build of Buck, because it required a native module (`bser.so`)
to be built as part of the build.

To work around this, I contributed a new pure Python `pybser.py` replacement for
the native code upstream:

https://github.com/facebook/watchman/pull/148

This diff integrates the new pure Python module with Buck and
pulls it into the PEX archive build.

Test Plan:
  ant && touch build/successful-build
  ./bin/buck build buck
  ./buck-out/gen/programs/buck.pex build buck

Confirmed in each log that BSER encoding was working.