Add a way to load platform-specific files from the workspace
Summary:
This makes it possible to have a `.win.{ext}`, `.linux.{ext}`, or `.macos.{ext}`
file loaded instead of the exact name provided.  This is useful if you have
something platform specific, such as paths, that you need to differentiate with.

Useful for https://github.com/facebook/buck/issues/319

Test Plan: Used in a later change