C/C++: propagate runtime deps from C/C++ library rules
Summary:
It's a common flow for a C/C++ library that includes code that uses
the output of another rule (e.g. calls a binary built by another C/C++
binary) to include a dependency on this rule to signify that Buck
should make this dependency available on-disk before running the code
in the library.  This diff makes sure to keep these as runtime deps in
C/C++ library rules so that they propagate up the dep chain to top-level
binary rules.

Test Plan: added a unittest for this case