Don't show warnings for superfluous R.txt entries in prebuilt aars
Summary:
Some R.txt files in prebuilt aars contain entries which are not defined
in the res directory, e.g. the R.txt in the google-play-services-base
aar contains the entries for all other play-services aars, but doesn't
define all of them in the res directory.

Because buck recently switched to using the resource entries from the
R.txt from a prebuilt aar, instead of generating them from the res
directory, this causes warnings such as: "Cannot find resource
'RDotTxtEntry{idType=int, type=attr, name=adSize, idValue=0x7f010000}'
in the uber R.txt.".

The user cannot do anything about these warnings, short of editing the
prebuilt aar, and they could be distracting from actual issues in the
project.

The warnings were originally useful when MiniAapt landed, but are no
longer generally useful.  Instead of showing the warnings to the user
send them to LOG.debug, in case they are needed for diagnosing issues.

In addition get rid of warnMissingResources.  As we are sending the
output to the debug log only, there is no need for the additional logic
to avoid showing such warnings to the user.

Test Plan:
add test
build project which includes the google-play-services-base