Avoid running ant when buck is already built
Summary:
When building buck using ant, no "successful-build" file is created in
the build directory.  Because buck checks this file to determine whether
to invoke ant or not to build buck from source when it is executed, ant
is unnecessarily invoked again when buck is run, when buck was build
directly using ant.

Instead of writing the file after the python wrapper executes ant, write
the file when ant compiles buck successfully.

Test Plan:
Build buck using ant, and then build a project using buck