Docs: Add missing space
Summary:
Soy templates don't add spaces on new lines when the previous line ends
with a html tag, or when the new line starts with one.  Some places in
the docs don't have a {sp} tag to force spaces on these lines.  Add
these tags where appropriate to improve the visualization of the docs.

In addition remove a stray </code> tag in the immutables documentation.

Closes https://github.com/facebook/buck/pull/436

Test Plan:
 docs/soyweb-local.sh