apple test: unzip xctool from build rule in build steps
Summary:
We were unpacking this in the test steps, which meant we were passing
a non-existent path to an external runner (and had to re-unpack it for
every test run).  This diff moves it to the build steps.

Test Plan: Ran with an external runner and verified xctool was unpacked