Make the DirCache not require the BuckEventBus.
Summary:
This makes the DirCache clean up when it is closed (and not when
it receives the BuildFinished event). It also makes the cache attempt
to trim itself every 50% of max size. Both of these changes are to
make it possible to serve the cache continuously without the risk
of its size growing unbounded.

Test Plan: CI