C/C++: refactor link blacklist to be a set of `BuildTarget`
Summary:
This is a precursor to some other refactoring of `NativeLinkables`
which aims to remove most of the use of `BuildRule`.  We currently
operate on `BuildRule`s and do filtering to get to `NativeLinkable`,
when we really should require the caller to pass in `NatveLinkable`
directly.  By switching this to `BuildTarget` it makes it a bit easier
to do this.

Test Plan: CI